requirements.txt
requests
tqdm
elasticsearch==8.12.1
sentence-transformers
transformers
pandas
pyyaml
